feat: onboarding refactor (#235)
## Overview

(I'm sure I'll miss some stuff)

- Each step has a route now, with an entrypoint `/welcome` and an exit
point `/completed`.
- Fully i18n'd, including (the fucking) `zod` and i18n's `t`. And
keeping TS somewhat happy.
- i18n slight refactor (no backend stuff).
- Onboarding specific i18n in separate JSON files.

### Routing

Red is commented out, not sure whether we want those redirects, or
whether they might affect the onboarding "reusability" for edit
user-profile.

![Untitled-2025-03-18-1107
excalidraw](https://github.com/user-attachments/assets/4b5c069b-e234-4154-a021-cb3ace2a05b1)


## Left to do

1. Character count for selected template. Shouldn't take long, but have
to "massage" the data a bit so we can do `templates[selectedTemplateId]`
as well as keep it in an array for the component. Alternatively, I
thought we might want to display some sort of counter, following the
idea of how it would look in a messaging app?

![Screenshot 2025-05-05 at 00 45
05](https://github.com/user-attachments/assets/0d899ea2-6227-47a7-94ef-af622a783eba)

## Bugs

0. Tailwind and Mantine weren't playing well after upgrading the former
to v4.
1. ~[All] If zod validation messages are shown, changing the global
language doesn't translate the errors (until the field is touched
again).~

## Things to improve

1. ~[Refactor: extract] Currently we're using `t:
TFunction<NotifycalI18nNamespaces, undefined>` in zod schema / resolver
for i18n in all steps. Would be nice if we could abstract that (and the
fix for the i18n validation message issue) into a hook that'd replace
and encapsulate `useForm`.~ Done: `useI18nForm`

## Doubts about implementations

1. Step order: AFAIK we had to use 2 Wizards before because we didn't
have access to the user-profile data. Now because the user data is
stored in zustand "globally", we don't need to retrieve the user-profile
data in order to send the reminder.

That being said, if that wasn't the case / constraint or reason it was
done like that, I'm happy to swap the step order. In the end the
`Confirm` step isn't that special anymore and could be in the middle
(and advance to the next step on successful request).

3. [ReminderType] There was a problem when changing template language
with a template id selected. After changing the language, from a user
PoV nothing is selected and the user can still go ahead. In reality,
there was a templateId saved, just not shown (because of the language
change).

```
language: es
templateId: formal-es-01

(template language change happens here, from es -> en)

language: en
templateId: formal-es-01
```

![Screenshot 2025-05-05 at 00 36
27](https://github.com/user-attachments/assets/60bd2b32-5127-4153-8580-9c662c3de159)
Template selected in Spanish

![Screenshot 2025-05-05 at 00 36
51](https://github.com/user-attachments/assets/c69b01e6-b72b-4deb-b748-f5764beba0a3)
After changing to English, the form lets us continue because the
previous template wasn't unselected. But for the user, nothing is
selected.

Solved it by resetting the templateId on language change, but I'm not
sure that's what we'd want?

4. Assess `onChange` form mode as [it might impact
performance](https://react-hook-form.com/docs/useform#mode)